+++ src/Components/UI/Organisms/Containers/index.ts
@@ -3,3 +3,4 @@ export { default as OrderStateListContainer } from "./OrderStateListContainer";
 export { default as SignupFormContainer } from "./SignupFormContainer";
 export { default as UserFormContainer } from "./UserFormContainer";
 export { default as StoreFormContainer } from "./StoreFormContainer";
+export { default as LoginFormContainer } from "./LoginFormContainer";
